refactor: clean up packet layout

+++ matchserver/include/ServerTypes.h
@@ -4,7 +4,7 @@
 using MatchId = uint32_t;
 using RequestId = uint32_t;
 using SessionId = uint32_t;
-using ServerTimeMs = uint64_t;
+using ServerTimeMs = uint32_t;
 
 constexpr uint16_t MAX_PACKET_SIZE{ 1024 };
 using RawBuffer = std::array<uint8_t, MAX_PACKET_SIZE>;
